% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculates the shortest distances between an initial node and all other nodes in an input data.frame}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm/}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A function argument. A data.frame}

\item{init_node}{The initialized vertex from which distances should be computed}

\item{v_v1}{A Vector of integers for the start positions}

\item{v_v2}{A Vector of integers for the stop positions}

\item{v_w}{A Vector of distances between the corresponding vertexes in vector v_v1 and v_v2}

\item{int_max_node}{The maximum vertex}

\item{v_answer}{A vector of distances between the initialized node and every other node}

\item{int_updated_node}{The current next node}

\item{v_connected_vortex_positions}{A vector of all connected positions to the current node}

\item{int_position}{An integer rutting over v_connected_vortex_positions}

\item{int_min_distance}{The shortest path to the nest vertex}

\item{v1}{A Vector of integers for the start positions}

\item{v2}{A Vector of integers for the stop positions}

\item{w}{A Vector of distances between the corresponding vertexes in vector v1 and v2}
}
\value{
The vector of shortest distances between the initialized node and every other node
}
\description{
Calculates the shortest distances between an initial node and all other nodes in an input data.frame
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
