% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculates the GCD for two integers}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm/}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A first integer}

\item{b}{A second integer}
}
\value{
The GCD
}
\description{
Calculates the GCD for two integers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
