% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeather.R
\name{getWeather}
\alias{getWeather}
\title{API-retrived weather near Norrkoping}
\usage{
getWeather(lon, lat)
}
\arguments{
\item{lon}{The longitude.}

\item{lat}{The latitude.}

\item{VariablesDf}{A data.frame that contains all variables of interest at a given day at an interval of 1 h.}

\item{CurrentDate}{The current date.}
}
\value{
The API that lets to retrieve information about weather near Norrkoping at longitude = {lon1, lon2, lo3, lo4} and at latitude = {lat1, lat2, lat3, lat4} at the current day.
}
\description{
API-retrived weather near Norrkoping
}
