% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{The graph from the Wikipedia page for the Dijkstra algorithm}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm/}
}
\usage{
wiki_graph
}
\description{
The graph from the Wikipedia page for the Dijkstra algorithm
}
\keyword{datasets}
