% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeather.R
\name{getWeather}
\alias{getWeather}
\title{API-retrived weather near Norrkoping}
\usage{
getWeather()
}
\arguments{
\item{location}{The longitude and latitude of the weather retrieved.}

\item{parameters}{The weather parameters obtained.}

\item{CurrentDate}{The current date.}
}
\value{
The API that lets to retrieve information about weather near Norrkoping at longitude = 16 and at latitude = 58 at the current day.
}
\description{
API-retrived weather near Norrkoping
}
\examples{
getWeather()
}
